+++ apps/fluent-tester/src/TestComponents/Badge/BadgeTest.tsx
@@ -38,7 +38,7 @@ export const BadgeTest: React.FunctionComponent = () => {
   const description =
     'A badge is an additional visual descriptor for UI elements. It can be used to denote numerical value, status or general information.\n Included in this spec are base properties for usage in all badge scenarios. Reference properties for other badge types i.e. size, layout, and style variations';
 
-  const spec = 'https://github.com/microsoft/fluentui-react-native/blob/main/packages/experimental/Badge/SPEC.md';
+  const spec = 'https://github.com/microsoft/fluentui-react-native/blob/main/packages/components/Badge/SPEC.md';
 
   return <Test name="Badge Test" description={description} spec={spec} sections={badgeSections} status={status} />;
 };