+++ packages/components/Badge/src/Badge.types.ts
@@ -24,6 +24,27 @@ export interface BadgeConfigurableProps {
   badgeColor?: BadgeColor;
 
   /**
+   * Set the text color
+   */
+  color?: ColorValue;
+
+  /*
+   * Source URL or name of the icon to show on the Badge.
+   */
+  icon?: IconSourcesType;
+
+  /**
+   * The icon color.
+   */
+  iconColor?: ColorValue;
+
+  /**
+   * Icon can be placed before or after Badge's content.
+   * @default before
+   */
+  iconPosition?: BadgeIconPosition;
+
+  /**
    * Badge position
    * @defaultvalue absolute
    */
@@ -49,19 +70,8 @@ export interface BadgeProps extends BadgeCoreProps, BadgeConfigurableProps {
    * @defaultvalue filled
    */
   appearance?: BadgeAppearance;
-
-  /*
-   * Source URL or name of the icon to show on the Badge.
-   */
-  icon?: IconSourcesType;
-
-  /**
-   * Icon can be placed before or after Badge's content.
-   * @default before
-   */
-  iconPosition?: BadgeIconPosition;
 }
-export interface BadgeCoreTokens extends LayoutTokens, FontTokens, IBorderTokens, IShadowTokens, IColorTokens {
+export interface BadgeCoreTokens extends LayoutTokens, FontTokens, IBorderTokens, IShadowTokens, Omit<IColorTokens, 'color'> {
   /**
    * Set the bottom edge of the Badge
    */
@@ -71,14 +81,12 @@ export interface BadgeCoreTokens extends LayoutTokens, FontTokens, IBorderTokens
    * The height of the Badge.
    */
   height?: number;
-  /**
-   * The icon color.
-   */
-  iconColor?: ColorValue;
+
   /**
    * The icon size.
    */
   iconSize?: number;
+
   /**
    * Set the left edge of the Badge
    */
@@ -93,7 +101,8 @@ export interface BadgeCoreTokens extends LayoutTokens, FontTokens, IBorderTokens
    * Set padding for text container when Badge contains
    * icons or images
    */
-  textPadding?: number;
+  textMargin?: number;
+
   /**
    * Set the top edge of the Badge
    */
@@ -122,12 +131,7 @@ export interface BadgeCoreTokens extends LayoutTokens, FontTokens, IBorderTokens
   square?: BadgeTokens;
 }
 export interface BadgeTokens extends BadgeCoreTokens, BadgeConfigurableProps {
-  /**
-   * The size of the icon.
-   */
-  iconSize?: number;
-
-  /**
+  /*
    * The weight of the lines used when drawing the icon.
    */
   iconWeight?: number;
@@ -160,7 +164,7 @@ export interface BadgeTokens extends BadgeCoreTokens, BadgeConfigurableProps {
 
 export interface BadgeSlotProps {
   root: IViewProps;
-  icon: IconProps;
+  icon?: IconProps;
   text: TextProps;
 }