+++ packages/components/Badge/src/BadgeColorTokens.ts
@@ -1,12 +1,15 @@
 import { Theme, TokenSettings } from '@fluentui-react-native/framework';
 import { globalTokens } from '@fluentui-react-native/theme-tokens';
 import { BadgeTokens } from './Badge.types';
-import { getFilledColorProps, getOutlineColorProps, getTintColorProps, getGhostColorProps, getWin32Props } from './colorHelper';
+import { getFilledColorProps, getOutlineColorProps, getTintColorProps, getGhostColorProps } from './colorHelper';
 
 export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
   ({
+    color: t.colors.neutralForegroundOnBrand,
+    backgroundColor: t.colors.brandBackgroundStatic,
+    borderColor: t.colors.transparentStroke,
     filled: {
-      borderColor: 'transparent',
+      borderColor: t.colors.transparentStroke,
     },
     outline: {
       backgroundColor: t.colors.transparentBackground,
@@ -22,11 +25,10 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             backgroundColor: t.colors.brandBackgroundStatic,
           },
           t,
-          getWin32Props,
         ),
       },
       outline: {
-        ...getOutlineColorProps({ color: t.colors.brandForeground1 }, t, getWin32Props),
+        ...getOutlineColorProps({ color: t.colors.brandForeground1 }, t),
       },
       tint: {
         ...getTintColorProps(
@@ -36,7 +38,6 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             borderColor: t.colors.brandStroke2,
           },
           t,
-          getWin32Props,
         ),
       },
       ghost: {
@@ -45,7 +46,6 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             color: t.colors.brandForeground1,
           },
           t,
-          getWin32Props,
         ),
       },
     },
@@ -56,11 +56,10 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             backgroundColor: globalTokens.color.red.primary,
           },
           t,
-          getWin32Props,
         ),
       },
       outline: {
-        ...getOutlineColorProps({ color: globalTokens.color.red.primary, colorDark: globalTokens.color.red.tint20 }, t, getWin32Props),
+        ...getOutlineColorProps({ color: globalTokens.color.red.primary, colorDark: globalTokens.color.red.tint20 }, t),
       },
       tint: {
         ...getTintColorProps(
@@ -73,7 +72,6 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             borderColorDark: globalTokens.color.red.shade30,
           },
           t,
-          getWin32Props,
         ),
       },
       ghost: {
@@ -83,7 +81,6 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             colorDark: globalTokens.color.red.tint20,
           },
           t,
-          getWin32Props,
         ),
       },
     },
@@ -94,15 +91,10 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             backgroundColor: globalTokens.color.darkOrange.primary,
           },
           t,
-          getWin32Props,
         ),
       },
       outline: {
-        ...getOutlineColorProps(
-          { color: globalTokens.color.darkOrange.primary, colorDark: globalTokens.color.orange.tint20 },
-          t,
-          getWin32Props,
-        ),
+        ...getOutlineColorProps({ color: globalTokens.color.darkOrange.primary, colorDark: globalTokens.color.orange.tint20 }, t),
       },
       tint: {
         ...getTintColorProps(
@@ -115,7 +107,6 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             borderColorDark: globalTokens.color.office.shade40,
           },
           t,
-          getWin32Props,
         ),
       },
       ghost: {
@@ -125,7 +116,6 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             colorDark: globalTokens.color.orange.tint20,
           },
           t,
-          getWin32Props,
         ),
       },
     },
@@ -137,15 +127,10 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             color: globalTokens.color.grey[14], // It should be neutralForegroundStatic1. It's hardcoded because the token doesn't exist right now
           },
           t,
-          getWin32Props,
         ),
       },
       outline: {
-        ...getOutlineColorProps(
-          { color: globalTokens.color.yellow.shade30, colorDark: globalTokens.color.yellow.tint40 },
-          t,
-          getWin32Props,
-        ),
+        ...getOutlineColorProps({ color: globalTokens.color.yellow.shade30, colorDark: globalTokens.color.yellow.tint40 }, t),
       },
       tint: {
         ...getTintColorProps(
@@ -158,7 +143,6 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             borderColorDark: globalTokens.color.yellow.shade30,
           },
           t,
-          getWin32Props,
         ),
       },
       ghost: {
@@ -168,7 +152,6 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             colorDark: globalTokens.color.yellow.tint40,
           },
           t,
-          getWin32Props,
         ),
       },
     },
@@ -179,11 +162,10 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             backgroundColor: globalTokens.color.green.primary,
           },
           t,
-          getWin32Props,
         ),
       },
       outline: {
-        ...getOutlineColorProps({ color: globalTokens.color.green.primary, colorDark: globalTokens.color.green.tint20 }, t, getWin32Props),
+        ...getOutlineColorProps({ color: globalTokens.color.green.primary, colorDark: globalTokens.color.green.tint20 }, t),
       },
       tint: {
         ...getTintColorProps(
@@ -196,7 +178,6 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             borderColorDark: globalTokens.color.green.shade30,
           },
           t,
-          getWin32Props,
         ),
       },
       ghost: {
@@ -206,7 +187,6 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             colorDark: globalTokens.color.green.tint20,
           },
           t,
-          getWin32Props,
         ),
       },
     },
@@ -218,11 +198,10 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             color: t.colors.neutralBackground1,
           },
           t,
-          getWin32Props,
         ),
       },
       outline: {
-        ...getOutlineColorProps({ color: t.colors.neutralForeground3, borderColor: t.colors.neutralStrokeAccessible }, t, getWin32Props),
+        ...getOutlineColorProps({ color: t.colors.neutralForeground3, borderColor: t.colors.neutralStrokeAccessible }, t),
       },
       tint: {
         ...getTintColorProps(
@@ -235,7 +214,6 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             borderColorDark: globalTokens.color.grey[68],
           },
           t,
-          getWin32Props,
         ),
       },
       ghost: {
@@ -244,7 +222,6 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             color: t.colors.neutralForeground1,
           },
           t,
-          getWin32Props,
         ),
       },
     },
@@ -258,11 +235,10 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             hcColor: t.colors.brandForeground1,
           },
           t,
-          getWin32Props,
         ),
       },
       outline: {
-        ...getOutlineColorProps({ color: t.colors.neutralForeground3, borderColor: t.colors.neutralStroke2 }, t, getWin32Props),
+        ...getOutlineColorProps({ color: t.colors.neutralForeground3, borderColor: t.colors.neutralStroke2 }, t),
       },
       tint: {
         ...getTintColorProps(
@@ -272,7 +248,6 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             borderColor: t.colors.neutralStroke2,
           },
           t,
-          getWin32Props,
         ),
       },
       ghost: {
@@ -281,7 +256,6 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             color: t.colors.neutralForeground3,
           },
           t,
-          getWin32Props,
         ),
       },
     },
@@ -295,11 +269,10 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             hcColor: t.colors.brandForeground1,
           },
           t,
-          getWin32Props,
         ),
       },
       outline: {
-        ...getOutlineColorProps({ color: t.colors.neutralForegroundOnBrand }, t, getWin32Props),
+        ...getOutlineColorProps({ color: t.colors.neutralForegroundOnBrand }, t),
       },
       tint: {
         ...getTintColorProps(
@@ -309,7 +282,6 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             borderColor: t.colors.neutralStroke2,
           },
           t,
-          getWin32Props,
         ),
       },
       ghost: {
@@ -318,7 +290,6 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
             color: t.colors.neutralForegroundOnBrand,
           },
           t,
-          getWin32Props,
         ),
       },
     },