+++ packages/components/Badge/CHANGELOG.md
@@ -1,8 +1,43 @@
 # Change Log - @fluentui-react-native/badge
 
-This log was last generated on Fri, 26 Aug 2022 18:15:16 GMT and should not be manually modified.
+This log was last generated on Tue, 30 Aug 2022 18:36:15 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.1.28
+
+Tue, 30 Aug 2022 18:36:15 GMT
+
+### Patches
+
+- Moved Badge to components folder. (vkozlova@microsoft.com)
+
+## 0.1.27
+
+Tue, 30 Aug 2022 08:26:56 GMT
+
+### Patches
+
+- Fixed customization (vkozlova@microsoft.com)
+
+## 0.1.26
+
+Mon, 29 Aug 2022 23:21:44 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.8.1
+- Bump @fluentui-react-native/icon to v0.13.5
+- Bump @fluentui-react-native/interactive-hooks to v0.18.2
+- Bump @fluentui-react-native/text to v0.14.1
+
+## 0.1.25
+
+Mon, 29 Aug 2022 12:39:37 GMT
+
+### Patches
+
+- Fixed showing content for tiny and extraSmall badge (vkozlova@microsoft.com)
 
 ## 0.1.24