+++ packages/experimental/Dropdown/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/dropdown
 
-This log was last generated on Fri, 26 Aug 2022 19:52:23 GMT and should not be manually modified.
+This log was last generated on Fri, 02 Sep 2022 05:01:01 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.1.2
+
+Fri, 02 Sep 2022 05:01:01 GMT
+
+### Patches
+
+- Skeletal Option component (ruaraki@microsoft.com)
+
+## 0.1.1
+
+Wed, 31 Aug 2022 00:27:59 GMT
+
+### Patches
+
+- Setting up compose framework for Dropdown (ruaraki@microsoft.com)
 
 ## 0.1.0